Move function to utils
Needed to shift around some code to avoid circular imports
